edit code inside directory '15. Inheritance/Lesson/starter/product' to follow lecture.
