+++ Module 2 - Object Oriented Programming/15. Inheritance/Lesson/starter/product/Pants.java
@@ -1,12 +1,9 @@
 package product;
 
-public class Pants {
+public class Pants extends Product {
 
     private int waist;
     private int length;
-    private double price;
-    private String color;
-    private String brand;
 
     public int getWaist() {
         return this.waist;
@@ -22,30 +19,6 @@ public class Pants {
 
     public void setLength(int length) {
         this.length = length;
-    }
-
-    public double getPrice() {
-        return this.price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-
-    public String getColor() {
-        return this.color;
-    }
-
-    public void setColor(String color) {
-        this.color = color;
-    }
-
-    public String getBrand() {
-        return this.brand;
-    }
-
-    public void setBrand(String brand) {
-        this.brand = brand;
     }
 
 }
